IdeaProjects/Restaurant_project/src/test/java/RestaurantTest.java
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import static org.junit.jupiter.api.Assertions.*;
import org.mockito.Mockito;
import org.mockito.junit.jupiter.MockitoExtension;
import org.mockito.junit.jupiter.MockitoSettings;
import org.mockito.quality.Strictness;
import org.mockito.Spy;

import static org.hamcrest.MatcherAssert.assertThat;

import java.time.LocalTime;
import java.time.format.DateTimeFormatter;
import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
@MockitoSettings(strictness = Strictness.LENIENT)

class RestaurantTest {
    Restaurant restaurant;
    //REFACTOR ALL THE REPEATED LINES OF CODE

    @BeforeEach
    public void setup(){
        LocalTime openingTime = LocalTime.parse("10:30:00");
        LocalTime closingTime = LocalTime.parse("22:00:00");
        restaurant = new Restaurant("Amelie's cafe", "Chennai", openingTime, closingTime);
    }

    //>>>>>>>>>>>>>>>>>>>>>>>>>OPEN/CLOSED<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
    //-------FOR THE 2 TESTS BELOW, YOU MAY USE THE CONCEPT OF MOCKING, IF YOU RUN INTO ANY TROUBLE
    @Test
    public void is_restaurant_open_should_return_true_if_time_is_between_opening_and_closing_time() {
        //WRITE UNIT TEST CASE HERE

        Restaurant restaurantSpy = Mockito.spy(restaurant);

        LocalTime fourhoursAfterCurrentTime = LocalTime.now().plusHours(4);
        Mockito.when(restaurantSpy.getCurrentTime()).thenReturn(fourhoursAfterCurrentTime);

        restaurant.isRestaurantOpen();

        assertEquals(true, restaurant.isRestaurantOpen());

    }

    @Test
    public void is_restaurant_open_should_return_false_if_time_is_outside_opening_and_closing_time() {
        //WRITE UNIT TEST CASE HERE

        Restaurant restaurantSpy = Mockito.spy(restaurant);

        LocalTime fourteenhoursAfterCurrentTime = LocalTime.now().plusHours(14);
        Mockito.when(restaurantSpy.getCurrentTime()).thenReturn(fourteenhoursAfterCurrentTime);

        restaurant.isRestaurantOpen();

        assertEquals(false, restaurant.isRestaurantOpen());

    }

    //<<<<<<<<<<<<<<<<<<<<<<<<<OPEN/CLOSED>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>


    //>>>>>>>>>>>>>>>>>>>>>>>>>>>MENU<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
    @Test
    public void adding_item_to_menu_should_increase_menu_size_by_1() {

        restaurant.addToMenu("Sweet corn soup", 119);
        restaurant.addToMenu("Vegetable lasagne", 269);

        int initialMenuSize = restaurant.getMenu().size();
        restaurant.addToMenu("Sizzling brownie", 319);
        assertEquals(initialMenuSize + 1, restaurant.getMenu().size());
    }

    @Test
    public void removing_item_from_menu_should_decrease_menu_size_by_1() throws itemNotFoundException {

        restaurant.addToMenu("Sweet corn soup", 119);
        restaurant.addToMenu("Vegetable lasagne", 269);

        int initialMenuSize = restaurant.getMenu().size();
        restaurant.removeFromMenu("Vegetable lasagne");
        assertEquals(initialMenuSize - 1, restaurant.getMenu().size());
    }

    @Test
    public void removing_item_that_does_not_exist_should_throw_exception() {

        restaurant.addToMenu("Sweet corn soup", 119);
        restaurant.addToMenu("Vegetable lasagne", 269);

        assertThrows(itemNotFoundException.class,
                () -> restaurant.removeFromMenu("French fries"));
    }
    //<<<<<<<<<<<<<<<<<<<<<<<MENU>>>>>>>>>>>>>>>>>>>>>>>>>>>>>


    //<<<<<<<<<<<<<<<<<<<<ESTIMATING>>>>>>>>>>>>>>>>>>>>>>>>>>
    //TDD

    //In the stub code, you need to add another method that returns the order value, given the name of the items in <String> format.

    //Estimate(order value)
    //Menu items
    //Item price


    @Test
    public void when_selecting_menu_item_names_should_display_order_total() {

        List<Item> itemList = new ArrayList<>();
        Item item1 = new Item("Sweet corn soup", 119);
        Item item2 = new Item("Vegetable lasagne", 269);

        itemList.add(item1);
        itemList.add(item2);

        int orderTotal = restaurant.selectedItemsOrderTotal(itemList);

        assertEquals(388, orderTotal);
    }

}
    //<<<<<<<<<<<<<<<<<<<<ESTIMATING>>>>>>>>>>>>>>>>>>>>>>>>>>